+++ preprocess.R
@@ -6,7 +6,6 @@ meta1 = sel.meta1
 meta2 = sel.meta2
 # mat1, meta1, mat2, meta2
 
-
 # Vars
 prefix = "data/LiScience2018/"
 mat = mat1
@@ -17,7 +16,6 @@ paste("Mat Dim:", dim(mat)[1], "x", dim(mat)[2], "  Meta Dim:", dim(meta)[1], "x
 dir.create(prefix)
 write.csv(mat, paste(prefix, "mat.csv", sep=""))
 write.csv(meta, paste(prefix, "meta.csv", sep=""))
-
 
 # Vars
 prefix = "data/GordonNature2021/"
@@ -31,17 +29,24 @@ write.csv(mat, paste(prefix, "mat.csv", sep=""))
 write.csv(meta, paste(prefix, "meta.csv", sep=""))
 
 
-
 ### Pseudo Data
 load("sc2sc.small.RData")
 meta1 = ps.meta1
 meta2 = ps.meta2
 # mat1, meta1, mat2, meta2
 
+# Filter cell types for visual
+for (i in 1:length(meta1$cellType)) {
+  if (!(meta1$cellType[i] %in% c("EN", "IN", "IPC", "RG", "Mural")))
+    meta1$cellType[i] = "Other"
+}
+for (i in 1:length(meta2$cellType)) {
+  if (!(meta2$cellType[i] %in% c("EN", "IN", "IPC", "RG", "Mural")))
+    meta2$cellType[i] = "Other"
+}
 
-### Misc
-write.csv(knn_ini, "corr.csv")
-
+# Misc
+# write.csv(knn_ini, "corr.csv")
 
 # Vars
 prefix = "data/NowakowskiScience2017/"
@@ -53,7 +58,6 @@ paste("Mat Dim:", dim(mat)[1], "x", dim(mat)[2], "  Meta Dim:", dim(meta)[1], "x
 dir.create(prefix)
 write.csv(mat, paste(prefix, "mat.csv", sep=""))
 write.csv(meta, paste(prefix, "meta.csv", sep=""))
-
 
 # Vars
 prefix = "data/KantonNature2019/"
@@ -67,14 +71,12 @@ write.csv(mat, paste(prefix, "mat.csv", sep=""))
 write.csv(meta, paste(prefix, "meta.csv", sep=""))
 
 
-
 ### All Data
 load("all.data.RData")
 # human.data1, human.data2, human.data3
 # human.meta1, human.meta2, human.meta3
 # org.data1, org.data2
 # org.meta1, org.meta2
-
 
 ### Works
 # # Vars
@@ -88,7 +90,6 @@ load("all.data.RData")
 # write.csv(mat, paste(prefix, "mat.csv", sep=""))
 # write.csv(meta, paste(prefix, "meta.csv", sep=""))
 
-
 # Vars
 prefix = "data/BireyNature2017/"
 mat = t(org.data1)
@@ -99,7 +100,6 @@ paste("Mat Dim:", dim(mat)[1], "x", dim(mat)[2], "  Meta Dim:", dim(meta)[1], "x
 dir.create(prefix)
 write.csv(mat, paste(prefix, "mat.csv", sep=""))
 write.csv(meta, paste(prefix, "meta.csv", sep=""))
-
 
 ### Doesn't work
 # Vars
@@ -113,7 +113,6 @@ dir.create(prefix)
 write.csv(mat, paste(prefix, "mat.csv", sep=""))
 write.csv(meta, paste(prefix, "meta.csv", sep=""))
 
-
 # Vars
 prefix = "data/BhaduriNature2020Brain/"
 mat = human.data3
@@ -124,7 +123,6 @@ paste("Mat Dim:", dim(mat)[1], "x", dim(mat)[2], "  Meta Dim:", dim(meta)[1], "x
 dir.create(prefix)
 write.csv(mat, paste(prefix, "mat.csv", sep=""))
 write.csv(meta, paste(prefix, "meta.csv", sep=""))
-
 
 # Vars
 prefix = "data/BhaduriNature2020Org/"